==> 01_download_metadata.ipynb <==
%load_ext autoreload
%autoreload 2

%matplotlib inline

# Descarga de Metadatos

Primer _Notebook_ a ejecutar para hacer una descarga de datos del servidor de Genomic Data Commons (GDC). Se obtiene información de todos los casos que cumplan el filtro indicado en el fichero de configuración llamado 'cases info'. Para cada uno de esos casos se obtienen los metadatos de las imágenes histólogicas disponibles y de ficheros de datos ómicos. Los datos se guardarán en la ruta indicada en la configuración (*data_path*). Los ficheros de salida son:


* cases.csv

* slides_metadata.csv

* rnaseq_metadata.csv

__Packages__

from gdc.metadata import get_rnaseq_metadata
from gdc.metadata import get_slides_metadata
from gdc.metadata import get_cases

import requests
import json
import yaml
import pandas as pd
import io
import os

pd.options.display.max_columns = 100

__Config__

with open('conf/user_conf.yaml', 'r') as f:
    conf = yaml.load(f)

## Cases Data

Descarga la información indicada para cada uno de los casos. Se ha seleccionado utilizar todos los datos del proyecto TCGA-PAAD (Pancreatic Adenocarcinoma).

conf['cases_info']

Los campos a obtener para cada uno de los pacientes también son configurables. Se puede ver una lista de los campos disponibles en la siguiente URL: https://docs.gdc.cancer.gov/API/Users_Guide/Appendix_A_Available_Fields/. Se puede obtener información demográfica del paciente, su consumo de alcohol y/o cigarrillos asi como datos del diagnóstico.

conf['fields']['cases']

La función *get_cases* obtiene un DataFrame con todos los casos y las columnas indicadas mediante una llamada a la API de GDC.

cases_df = get_cases(conf['cases_info'], conf['fields']['cases'])

Guarda la salida en un CSV

cases_df.to_csv(os.path.join(conf['data_path'], 'cases.csv'), sep='|', index=False)

Muestra de la salida:

cases_df.sample(5)

## Files Metadata

### Slides Metadata

A continuación se obienen los metadatos de las imágenes histológicas. Además del filtro de pacientes que se aplica en los casos también se puede filtrar el tipo de _slide_. Las opciones diponibles son:

* _Tissue Slide_: muestras almezandas mediante congelación.

* _Diagnostic Slide_: tejidos embebidos en parafina y fijados con formalina (FFPE).

Las muestras FFPE son las más utlizadas por los médicos para el diagnóstico y el tejido se converva mejor, sin embargo los químicos en los que se almacenan afectan a los ácididos nucleicos, no siendo válidos para anális de ADN o RNA. 

Las muestras congeladas son más baratas de almacenar y suelen ser utilizar después de la cirujía para determinar si los bordes del tumor están limpios, es decir, si el tumor ha sido completamente eliminado. Por tanto sobre estas muestras se recogen datos sobre el porcentaje de las células afectadas por el tumor. Otra ventaja de estas imágenes es que es muy común que se haya recogido también su información genética, lo que permitiría un análisis de correlación entre la morfología del tejido tumoral y la experesión genética.

En este caso se han seleccionado ambos tipos para mostrar la diferencia de los datos disonibles pero para el proyecto se utilarán únicamente tejidos congelados por los motivos expuestos.

conf['slide_types']

slides_df = get_slides_metadata(conf['cases_info'], conf['fields']['slides'], experimental_strategies=conf['slide_types'])

slides_df.to_csv(os.path.join(conf['data_path'], 'slides_metadata.csv'), sep='|', index=False)

Muestra de tejidos congelados

slides_df[slides_df['experimental_strategy'] == 'Tissue Side'].sample(3)

Muestra de tejidos FFPE

Se observa que no las columnas *percent_normal_cells, percent_stromal_cells, percent_tumor_cells, percent_tumor_nuclei* están siempre a nulo ya que no se recogen estos datos.

slides_df[slides_df['experimental_strategy'] == 'Diagnostic Slide'].sample(3)

El 97% de las muestras FFPE pertenecen a tejidos tumorales.

slides_df.groupby(['experimental_strategy', 'sample_type']).size()

### RNA-Seq Metadata

Por último se obtienen los nombres de los ficheros de expresión genética, en ese caso también se puede seleccionar qué tipo de fichero queremos, las opciones son:

* HTSeq - Counts

* HTSeq - FPKM

* HTSeq - FPKM-UQ

El primero tiene la cuenta en bruto de las lecturas mapeadas a cada gen. Los otros dos están normalizados respecto al número de cuentas, la diferencia entre ellos es que _UQ_ utiliza el percentil 75% mientras que el otro la cuenta total.

conf['rnaseq_types']

rnaseq_df = get_rnaseq_metadata(conf['cases_info'], conf['fields']['rnaseq'], workflow_types=conf['rnaseq_types'])

rnaseq_df.to_csv(os.path.join(conf['data_path'], 'rnaseq_metadata.csv'), sep='|', index=False)

Observar que para cada una de las muestras se obtienen los 3 ficheros de RNA-Seq.

rnaseq_df.sort_values('sample_id').head(6)
