==> 02_download_files.ipynb <==
%load_ext autoreload
%autoreload 2

# Descarga de Ficheros

En este paso se descargan los ficheros de imágenes histológicas y de expresión genénica a partir de la información obtenida en el apartado anterior.

__Packages__

from gdc.download import api_download_iterative
from gdc.utils import gunzip

import requests
import yaml
import json
import os
import re
import pandas as pd

__Config__

with open('conf/user_conf.yaml', 'r') as f:
    conf = yaml.load(f)

__Paths__

slides_metadata_file = os.path.join(conf['data_path'], 'slides_metadata.csv')
rnaseq_metadata_file = os.path.join(conf['data_path'], 'rnaseq_metadata.csv')

slides_path = os.path.join(conf['data_path'], 'slides', 'svs')
if not os.path.exists(slides_path):
    os.mkdir(slides_path)

rnaseq_path = os.path.join(conf['data_path'], 'rnaseq')

if not os.path.exists(rnaseq_path):
    os.mkdir(rnaseq_path)

## Lectura de Metadatos

Obtiene el _id_ de los ficheros a descargar a partir de los metadatos descargados anteriormente. 

slides_df = pd.read_csv(slides_metadata_file, sep='|')
slides_df = slides_df[['file_name', 'file_id', 'file_size', 'sample_id', 'experimental_strategy']]
slides_df.head(3)

rnaseq_df = pd.read_csv(rnaseq_metadata_file, sep='|')
rnaseq_df = rnaseq_df[['file_name', 'file_id', 'file_size', 'sample_id', 'workflow_type']]
rnaseq_df.head(3)

## Slides

Las imágenes histológicas están en formato SVS de alta resolución, por tanto la descarga será lenta y ocuparán mucho en disco.

summary = slides_df.groupby('experimental_strategy').agg({'file_name': 'size', 'file_size': 'sum'})
summary = summary.rename(columns={'file_name': 'count', 'file_size': 'total_size (gb)'})
summary['total_size (gb)'] = round(summary['total_size (gb)'] / 1000, 2)

summary

Dependiendo del estudio que se desee realizar se puede que únicamente queramos utilizar aquellas imágenes que tengan su correspondiente información genénica. En ese caso se filtra el DataFrame de láminas haciendo que la muestra esté también en el DataFrame de RNA-SEQ.

rna_seq_samples = rnaseq_df['sample_id'].unique()
slides_df = slides_df[slides_df['sample_id'].isin(rna_seq_samples)]

summary = slides_df.groupby('experimental_strategy').agg({'file_name': 'size', 'file_size': 'sum'})
summary = summary.rename(columns={'file_name': 'count', 'file_size': 'total_size (gb)'})
summary['total_size (gb)'] = round(summary['total_size (gb)'] / 1000, 2)

summary

La función _api_download_iterative_ descarga todas las imágenes en la ruta indicada, con el parámetro _multiprocess_ es posible realizar descargas en paralelo, se recomienda utiliar en número de cores del procesador como número de descargas en paralelo. Se incluyen barras de progreso para ver el estado de las descargas.

api_download_iterative(slides_df, slides_path, multiprocess=8)

## RNA-Seq

Los ficheros de RNA-Seq al ser de texto plano ocuparán mucho menos y la descarga es rápida.

summary = rnaseq_df.groupby('workflow_type').agg({'file_name': 'size', 'file_size': 'sum'})
summary = summary.rename(columns={'file_name': 'count', 'file_size': 'total_size (gb)'})
summary['total_size (gb)'] = round(summary['total_size (gb)'] / 1000, 2)

summary

De nuevo se pueden filtrar aquellas muestras que estén pareadas con datos histológicos, en ese caso son todas.

slides_samples = slides_df['sample_id'].unique()
rnaseq_df = rnaseq_df[rnaseq_df['sample_id'].isin(slides_samples)]

summary = rnaseq_df.groupby('workflow_type').agg({'file_name': 'size', 'file_size': 'sum'})
summary = summary.rename(columns={'file_name': 'count', 'file_size': 'total_size (gb)'})
summary['total_size (gb)'] = round(summary['total_size (gb)'] / 1000, 2)

summary

files = api_download_iterative(rnaseq_df, rnaseq_path, multiprocess=8)

Puesto que las imágenes viene comprimidas en formato _GNU ZIP_ será necesario descomprimirlas. Este paso itera sobre todas los ficheros en la ruta de descarga y descomprime aquellos con extensión *.gz*.

for file_name in files:
    
    source_filepath = os.path.join(rnaseq_path, file_name)
    dest_filepath = re.sub('\.gz$', '', source_filepath)
    
    gunzip(source_filepath, dest_filepath)
    os.remove(source_filepath)
