==> 03_WSI_patching.ipynb <==
%load_ext autoreload
%autoreload 2

%matplotlib inline

# Obtención de patches de imágenes WSI (Whole Slide Images)

Las imágenes en formato _svs_ son muy útiles para los médicos en ayuda al diagnóstico y para análisis individualizados debido a su alta resolución pero este formato no es válido como entrada a una red neuronal. 

Por un lado es necesaria la conversión a un formato de una única capa como _png_ o _jpg_ en lugar del formato piramidal _svs_. 

El otro problema es la resolución, de miles por miles de pixeles, una resolución de 1024x1024 ya significa que la capa de entrada de la red neuronal tiene más de un millón de neuronas; en en el caso de imágenes en blanco y negro, a color habría que multiplicar por 3 al tener 3 canales. Para solucionar este problema se divive cada imagen en *patches* (parches), cuadrados más pequeños sobre imagen original. 

En este Notebook se realizará el parcheado de cada una de las imágenes descargadas anteriormente, guardando en formato .png cada uno de los *patches* obtenidos.

Para entender bien este concepto de _patching_ sobre una imagen SVS se recomienda ver el ejemplo en el notebook *wsi_patching_example.ipynb* o su correspondiente html *wsi_patching_example.html*

**Librerías**

import yaml
import os
import openslide
import pandas as pd
pd.set_option('display.max_colwidth', -1)

import matplotlib.image as mpimg
from matplotlib import pyplot as plt

try:
    get_ipython()
    from tqdm import tqdm_notebook as tqdm
except:
    from tqdm import tqdm

from wsi.slide import thumbnail
from wsi.patch import patch_slides

**Config**

with open('conf/user_conf.yaml', 'r') as f:
    conf = yaml.load(f)

__Creación de rutas__

Se crean los directorios para guardar las imágenes de salida.

slides_path = os.path.join(conf['data_path'], 'slides', 'svs')

patches_path = os.path.join(conf['data_path'], 'slides', 'patches')

if not os.path.exists(patches_path):
    os.mkdir(patches_path)

thumbnails_path = os.path.join(conf['data_path'], 'slides', 'thumbnail')

if not os.path.exists(thumbnails_path):
    os.mkdir(thumbnails_path)

**Lectura de DataFrame de slides**

Se lee el DataFrame de slides para obtener el nombre de todas las imágenes descargadas. Se filtra por imágenes contenidas en el directorio por si alguna no hubiera podido ser descargada.

slides_df = pd.read_csv(os.path.join(conf['data_path'], 'slides_metadata.csv'), sep='|')
slides_df = slides_df[slides_df['file_name'].isin(os.listdir(slides_path))]

slides_df.head(2)

slides_df.groupby(['experimental_strategy', 'sample_type']).size()

## Guarda miniaturas

Para facilitar el trabajo cuando se quiera visualizar una imagen al completo se guardará en formato .png una una imágen en baja resolución de la imagen SVS original. De esta manera no será necesario utilizar herramientas de lectura de imágenes SVS para hacerse una idea de la estructura completa de una lámina.

La resolución de esta imagen de sale se selecciona en el fichero de configuración bajo el parámetro *thumbnail_size*.

conf['wsi']['thumbnail_size']

for file in tqdm(slides_df['file_name'].values, unit='file'):
    
    os_img = openslide.open_slide(os.path.join(slides_path, file))
    img = thumbnail(os_img, max_size=conf['wsi']['thumbnail_size'])
    
    img.save(os.path.join(thumbnails_path, file.replace('.svs', '.png')))

## Patching

Finalmente se realiza el parcheado de cada una de las imágenes. La función *patch_slides* recibe una lista de nombres de ficheros y almacena todos los *patches* generados en un directorio. Los parámetros que se pueden modificar sobre el parcheado son los siguienes:

* _patch_size_: tamaño en píxeles de cada uno de los cuadrados.

* _magnification_: nivel de aumento microscópico sobre el que realizar el parcheado. A mayor aumento mayor resolución de la imágen original y por tanto mayor número de parches.

* *white_pixel_threshold*: es el umbral del porcentaje máximo de píxeles blancos que puede tener un parche para ser aceptado (ver notebook the ejemplo).

* sampling: tasa de muestreo de los parches. Cuando se utilizan niveles de aumento elevados el número de parches es muy alto y se recomienda hacer un muestreo.

conf['wsi']

slide_files = slides_df['file_name'].map(lambda x: os.path.join(slides_path, x))
slide_files.iloc[:5]

results = patch_slides(slide_files, patches_path, conf['wsi']['patch_size'],  conf['wsi']['magnification'], 
                       conf['wsi']['white_pixel_threshold'], conf['wsi']['sampling'])

Además de guardar los parches la función también devolverá un DataFrame con un resumen de los resultados para cada imágen, indicando el total de parches generados y cuántos de ellos fueron guardados.

results.head(5)

results['total_patches'].sum()

results['saved_patches'].sum()

round(results['perc_saved_patches'].mean(), 2)

results.to_csv(os.path.join(conf['data_path'], 'patching_results.csv'), sep='|', index=False)
